Move static code used at runtime by aot compiled modules to a separate package
