+++ aot/src/main/java/com/dylibso/chicory/aot/AotEmitters.java
@@ -1,36 +1,36 @@
 package com.dylibso.chicory.aot;
 
-import static com.dylibso.chicory.aot.AotMethods.CHECK_INTERRUPTION;
-import static com.dylibso.chicory.aot.AotMethods.INSTANCE_READ_GLOBAL;
-import static com.dylibso.chicory.aot.AotMethods.INSTANCE_SET_ELEMENT;
-import static com.dylibso.chicory.aot.AotMethods.MEMORY_COPY;
-import static com.dylibso.chicory.aot.AotMethods.MEMORY_DROP;
-import static com.dylibso.chicory.aot.AotMethods.MEMORY_FILL;
-import static com.dylibso.chicory.aot.AotMethods.MEMORY_GROW;
-import static com.dylibso.chicory.aot.AotMethods.MEMORY_INIT;
-import static com.dylibso.chicory.aot.AotMethods.MEMORY_PAGES;
-import static com.dylibso.chicory.aot.AotMethods.MEMORY_READ_BYTE;
-import static com.dylibso.chicory.aot.AotMethods.MEMORY_READ_DOUBLE;
-import static com.dylibso.chicory.aot.AotMethods.MEMORY_READ_FLOAT;
-import static com.dylibso.chicory.aot.AotMethods.MEMORY_READ_INT;
-import static com.dylibso.chicory.aot.AotMethods.MEMORY_READ_LONG;
-import static com.dylibso.chicory.aot.AotMethods.MEMORY_READ_SHORT;
-import static com.dylibso.chicory.aot.AotMethods.MEMORY_WRITE_BYTE;
-import static com.dylibso.chicory.aot.AotMethods.MEMORY_WRITE_DOUBLE;
-import static com.dylibso.chicory.aot.AotMethods.MEMORY_WRITE_FLOAT;
-import static com.dylibso.chicory.aot.AotMethods.MEMORY_WRITE_INT;
-import static com.dylibso.chicory.aot.AotMethods.MEMORY_WRITE_LONG;
-import static com.dylibso.chicory.aot.AotMethods.MEMORY_WRITE_SHORT;
-import static com.dylibso.chicory.aot.AotMethods.REF_IS_NULL;
-import static com.dylibso.chicory.aot.AotMethods.TABLE_COPY;
-import static com.dylibso.chicory.aot.AotMethods.TABLE_FILL;
-import static com.dylibso.chicory.aot.AotMethods.TABLE_GET;
-import static com.dylibso.chicory.aot.AotMethods.TABLE_GROW;
-import static com.dylibso.chicory.aot.AotMethods.TABLE_INIT;
-import static com.dylibso.chicory.aot.AotMethods.TABLE_SET;
-import static com.dylibso.chicory.aot.AotMethods.TABLE_SIZE;
-import static com.dylibso.chicory.aot.AotMethods.THROW_OUT_OF_BOUNDS_MEMORY_ACCESS;
-import static com.dylibso.chicory.aot.AotMethods.WRITE_GLOBAL;
+import static com.dylibso.chicory.aot.AotMethodRefs.CHECK_INTERRUPTION;
+import static com.dylibso.chicory.aot.AotMethodRefs.INSTANCE_READ_GLOBAL;
+import static com.dylibso.chicory.aot.AotMethodRefs.INSTANCE_SET_ELEMENT;
+import static com.dylibso.chicory.aot.AotMethodRefs.MEMORY_COPY;
+import static com.dylibso.chicory.aot.AotMethodRefs.MEMORY_DROP;
+import static com.dylibso.chicory.aot.AotMethodRefs.MEMORY_FILL;
+import static com.dylibso.chicory.aot.AotMethodRefs.MEMORY_GROW;
+import static com.dylibso.chicory.aot.AotMethodRefs.MEMORY_INIT;
+import static com.dylibso.chicory.aot.AotMethodRefs.MEMORY_PAGES;
+import static com.dylibso.chicory.aot.AotMethodRefs.MEMORY_READ_BYTE;
+import static com.dylibso.chicory.aot.AotMethodRefs.MEMORY_READ_DOUBLE;
+import static com.dylibso.chicory.aot.AotMethodRefs.MEMORY_READ_FLOAT;
+import static com.dylibso.chicory.aot.AotMethodRefs.MEMORY_READ_INT;
+import static com.dylibso.chicory.aot.AotMethodRefs.MEMORY_READ_LONG;
+import static com.dylibso.chicory.aot.AotMethodRefs.MEMORY_READ_SHORT;
+import static com.dylibso.chicory.aot.AotMethodRefs.MEMORY_WRITE_BYTE;
+import static com.dylibso.chicory.aot.AotMethodRefs.MEMORY_WRITE_DOUBLE;
+import static com.dylibso.chicory.aot.AotMethodRefs.MEMORY_WRITE_FLOAT;
+import static com.dylibso.chicory.aot.AotMethodRefs.MEMORY_WRITE_INT;
+import static com.dylibso.chicory.aot.AotMethodRefs.MEMORY_WRITE_LONG;
+import static com.dylibso.chicory.aot.AotMethodRefs.MEMORY_WRITE_SHORT;
+import static com.dylibso.chicory.aot.AotMethodRefs.REF_IS_NULL;
+import static com.dylibso.chicory.aot.AotMethodRefs.TABLE_COPY;
+import static com.dylibso.chicory.aot.AotMethodRefs.TABLE_FILL;
+import static com.dylibso.chicory.aot.AotMethodRefs.TABLE_GET;
+import static com.dylibso.chicory.aot.AotMethodRefs.TABLE_GROW;
+import static com.dylibso.chicory.aot.AotMethodRefs.TABLE_INIT;
+import static com.dylibso.chicory.aot.AotMethodRefs.TABLE_SET;
+import static com.dylibso.chicory.aot.AotMethodRefs.TABLE_SIZE;
+import static com.dylibso.chicory.aot.AotMethodRefs.THROW_OUT_OF_BOUNDS_MEMORY_ACCESS;
+import static com.dylibso.chicory.aot.AotMethodRefs.WRITE_GLOBAL;
 import static com.dylibso.chicory.aot.AotUtil.StackSize;
 import static com.dylibso.chicory.aot.AotUtil.callIndirectMethodName;
 import static com.dylibso.chicory.aot.AotUtil.callIndirectMethodType;
