+++ aot/src/test/resources/com/dylibso/chicory/approvals/ApprovalTest.verifyMemory.approved.txt
@@ -17,11 +17,11 @@ public final class com/dylibso/chicory/$gen/CompiledModule {
     ILOAD 0
     ICONST_0
     ALOAD 1
-    INVOKESTATIC com/dylibso/chicory/aot/AotMethods.memoryWriteInt (IIILcom/dylibso/chicory/runtime/Memory;)V
+    INVOKESTATIC com/dylibso/chicory/aot/runtime/AotMethods.memoryWriteInt (IIILcom/dylibso/chicory/runtime/Memory;)V
     ICONST_0
     ICONST_0
     ALOAD 1
-    INVOKESTATIC com/dylibso/chicory/aot/AotMethods.memoryReadInt (IILcom/dylibso/chicory/runtime/Memory;)I
+    INVOKESTATIC com/dylibso/chicory/aot/runtime/AotMethods.memoryReadInt (IILcom/dylibso/chicory/runtime/Memory;)I
     IRETURN
    L0
     ATHROW
@@ -32,25 +32,25 @@ public final class com/dylibso/chicory/$gen/CompiledModule {
     LLOAD 0
     ICONST_0
     ALOAD 2
-    INVOKESTATIC com/dylibso/chicory/aot/AotMethods.memoryWriteLong (IJILcom/dylibso/chicory/runtime/Memory;)V
+    INVOKESTATIC com/dylibso/chicory/aot/runtime/AotMethods.memoryWriteLong (IJILcom/dylibso/chicory/runtime/Memory;)V
     ICONST_0
     ICONST_0
     ALOAD 2
-    INVOKESTATIC com/dylibso/chicory/aot/AotMethods.memoryReadLong (IILcom/dylibso/chicory/runtime/Memory;)J
+    INVOKESTATIC com/dylibso/chicory/aot/runtime/AotMethods.memoryReadLong (IILcom/dylibso/chicory/runtime/Memory;)J
     LRETURN
    L0
     ATHROW
 
   // access flags 0x9
   public static call_indirect_0(IIILcom/dylibso/chicory/runtime/Memory;Lcom/dylibso/chicory/runtime/Instance;)I
-    INVOKESTATIC com/dylibso/chicory/aot/AotMethods.checkInterruption ()V
+    INVOKESTATIC com/dylibso/chicory/aot/runtime/AotMethods.checkInterruption ()V
     ALOAD 4
     ILOAD 2
     INVOKEVIRTUAL com/dylibso/chicory/runtime/Instance.table (I)Lcom/dylibso/chicory/runtime/TableInstance;
     ASTORE 5
     ALOAD 5
     ILOAD 1
-    INVOKESTATIC com/dylibso/chicory/aot/AotMethods.tableRef (Lcom/dylibso/chicory/runtime/TableInstance;I)I
+    INVOKESTATIC com/dylibso/chicory/aot/runtime/AotMethods.tableRef (Lcom/dylibso/chicory/runtime/TableInstance;I)I
     ISTORE 6
     ALOAD 5
     ILOAD 1
@@ -73,7 +73,7 @@ public final class com/dylibso/chicory/$gen/CompiledModule {
     INVOKESTATIC com/dylibso/chicory/$gen/CompiledModule.func_0 (ILcom/dylibso/chicory/runtime/Memory;Lcom/dylibso/chicory/runtime/Instance;)I
     GOTO L4
    L3
-    INVOKESTATIC com/dylibso/chicory/aot/AotMethods.throwIndirectCallTypeMismatch ()Ljava/lang/RuntimeException;
+    INVOKESTATIC com/dylibso/chicory/aot/runtime/AotMethods.throwIndirectCallTypeMismatch ()Ljava/lang/RuntimeException;
     ATHROW
    L4
     IRETURN
@@ -88,7 +88,7 @@ public final class com/dylibso/chicory/$gen/CompiledModule {
     ICONST_0
     ILOAD 6
     ALOAD 7
-    INVOKESTATIC com/dylibso/chicory/aot/AotMethods.callIndirect ([JIILcom/dylibso/chicory/runtime/Instance;)[J
+    INVOKESTATIC com/dylibso/chicory/aot/runtime/AotMethods.callIndirect ([JIILcom/dylibso/chicory/runtime/Instance;)[J
     ICONST_0
     LALOAD
     L2I
@@ -96,14 +96,14 @@ public final class com/dylibso/chicory/$gen/CompiledModule {
 
   // access flags 0x9
   public static call_indirect_1(JIILcom/dylibso/chicory/runtime/Memory;Lcom/dylibso/chicory/runtime/Instance;)J
-    INVOKESTATIC com/dylibso/chicory/aot/AotMethods.checkInterruption ()V
+    INVOKESTATIC com/dylibso/chicory/aot/runtime/AotMethods.checkInterruption ()V
     ALOAD 5
     ILOAD 3
     INVOKEVIRTUAL com/dylibso/chicory/runtime/Instance.table (I)Lcom/dylibso/chicory/runtime/TableInstance;
     ASTORE 6
     ALOAD 6
     ILOAD 2
-    INVOKESTATIC com/dylibso/chicory/aot/AotMethods.tableRef (Lcom/dylibso/chicory/runtime/TableInstance;I)I
+    INVOKESTATIC com/dylibso/chicory/aot/runtime/AotMethods.tableRef (Lcom/dylibso/chicory/runtime/TableInstance;I)I
     ISTORE 7
     ALOAD 6
     ILOAD 2
@@ -126,7 +126,7 @@ public final class com/dylibso/chicory/$gen/CompiledModule {
     INVOKESTATIC com/dylibso/chicory/$gen/CompiledModule.func_1 (JLcom/dylibso/chicory/runtime/Memory;Lcom/dylibso/chicory/runtime/Instance;)J
     GOTO L4
    L3
-    INVOKESTATIC com/dylibso/chicory/aot/AotMethods.throwIndirectCallTypeMismatch ()Ljava/lang/RuntimeException;
+    INVOKESTATIC com/dylibso/chicory/aot/runtime/AotMethods.throwIndirectCallTypeMismatch ()Ljava/lang/RuntimeException;
     ATHROW
    L4
     LRETURN
@@ -140,7 +140,7 @@ public final class com/dylibso/chicory/$gen/CompiledModule {
     ICONST_1
     ILOAD 7
     ALOAD 8
-    INVOKESTATIC com/dylibso/chicory/aot/AotMethods.callIndirect ([JIILcom/dylibso/chicory/runtime/Instance;)[J
+    INVOKESTATIC com/dylibso/chicory/aot/runtime/AotMethods.callIndirect ([JIILcom/dylibso/chicory/runtime/Instance;)[J
     ICONST_0
     LALOAD
     LRETURN
