+++ pom.xml
@@ -35,6 +35,7 @@
 
   <modules>
     <module>aot</module>
+    <module>aot-runtime</module>
     <module>aot-tests</module>
     <module>bom</module>
     <module>cli</module>
@@ -136,6 +137,11 @@
       <dependency>
         <groupId>com.dylibso.chicory</groupId>
         <artifactId>aot</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>com.dylibso.chicory</groupId>
+        <artifactId>aot-runtime</artifactId>
         <version>${project.version}</version>
       </dependency>
       <dependency>
